test: specify types for indices

+++ tests/test_1072_sort.py
@@ -348,10 +348,10 @@ def test_emptyarray_sort():
 
 def test_listarray_sort():
     v2_array = ak.contents.listarray.ListArray(
-        ak.index.Index(np.array([4, 100, 1])),
-        ak.index.Index(np.array([7, 100, 3, 200])),
+        ak.index.Index(np.array([4, 100, 1], dtype=np.int32)),
+        ak.index.Index(np.array([7, 100, 3, 200], dtype=np.int32)),
         ak.contents.numpyarray.NumpyArray(
-            np.array([6.6, 4.4, 5.5, 7.7, 3.3, 2.2, 1.1, 8.8])
+            np.array([6.6, 4.4, 5.5, 7.7, 3.3, 2.2, 1.1, 8.8], dtype=np.float64)
         ),
     )
